requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Deep Learning
torch==2.1.0
torchvision==0.16.0

# Computer Vision
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78

# Scientific Computing
numpy==1.24.3
scipy==1.11.4

# Utilities
Pillow==10.1.0
aiofiles==23.2.1

# Jupyter Notebook
jupyter==1.0.0
notebook==7.0.6
ipywidgets==8.1.1
matplotlib==3.8.2
seaborn==0.13.0
